move deposits and notifications
