+++ packages/core-sdk/src/lib/main.ts
@@ -16,6 +16,8 @@ import { KrayonTradeSDK } from './trade/trade-sdk';
 import { KrayonAPIClient } from './api-client';
 import { SDKReadyStatus } from './consts/enums';
 import { KrayonGasStationSDK } from './gas-station/gas-station-sdk';
+import { KrayonDepositSDK } from './deposit/deposit-sdk';
+import { KrayonNotificationSDK } from './notification/notification-sdk';
 
 export interface KrayonSDKConfig {
   token: string;
@@ -43,6 +45,8 @@ export class KrayonSDK {
   whitelist: KrayonWhitelistnSDK;
   gasStation: KrayonGasStationSDK;
   walletGroup: KrayonWalletGroupSDK;
+  deposit: KrayonDepositSDK;
+  notification: KrayonNotificationSDK;
   // For now, don't use the KrayonAPIClient, but rather, the singleton one
   // Easy to swap it later on
   // defaultApiClient: KrayonAPIClient;
@@ -81,7 +85,8 @@ export class KrayonSDK {
     this.user = new KrayonUserSDK({ apiClient });
     this.whitelist = new KrayonWhitelistnSDK({ apiClient });
     this.gasStation = new KrayonGasStationSDK({ apiClient });
-
+    this.deposit = new KrayonDepositSDK({ apiClient });
+    this.notification = new KrayonNotificationSDK({ apiClient });
     // Lastly, initialize all the SDKs that are (currently) dependent on the organization ID
     // We should eventually refactor these deps away, and simply be able to initialize everything here
     // But for now, this is the easiest way to do it, and we simply rerun this function in the
