+++ packages/core-sdk/src/lib/organization/organization-sdk.ts
@@ -32,7 +32,7 @@ import { GasStationDto } from '../types';
 import {
   MerchantDepositsResponse,
   MerchantNotificationResponse,
-} from '../types/merchant';
+} from '../types/deposit';
 
 export type OrganizationAssetFilter = Partial<Asset> & PaginationRequest;
 
@@ -308,21 +308,6 @@ export class KrayonOrganizationSDK {
         signal: abortSignal,
       }
     );
-  }
-
-  // MERCHANT - consider moving to merchant SDK
-  getDeposits(extraParams?: KrayonAPICommonOptions) {
-    const { abortSignal } = extraParams || {};
-    return this.apiClient.get<MerchantDepositsResponse>(`/deposits`, {
-      signal: abortSignal,
-    });
-  }
-
-  getNotifications(extraParams?: KrayonAPICommonOptions) {
-    const { abortSignal } = extraParams || {};
-    return this.apiClient.get<MerchantNotificationResponse>(`/notifications`, {
-      signal: abortSignal,
-    });
   }
 
   companyVerification(
