+++ packages/core-sdk/src/lib/types/deposit.ts
@@ -25,25 +25,4 @@ export interface MerchantDeposit {
   type: string;
 }
 
-export interface MerchantNotificationDeposit {
-  id: string;
-  organization: string;
-  wallet: string;
-  status: string;
-  amount: number;
-  currency: string;
-  symbol: string;
-  webhook_url: string;
-  type: string;
-}
-
-export interface MerchantNotification {
-  id: string;
-  created_at: string;
-  delivery_status: DepositStatus;
-  delivery_error_message: string;
-  related_object_details: MerchantNotificationDeposit | any;
-}
-
 export type MerchantDepositsResponse = Pageable<MerchantDeposit>;
-export type MerchantNotificationResponse = Pageable<MerchantNotification>;
