+++ packages/core-sdk/src/lib/types/index.ts
@@ -19,4 +19,5 @@ export * from './transfer';
 export * from './user';
 export * from './util';
 export * from './wallet';
-export * from './merchant';
+export * from './deposit';
+export * from './notification';
